fix frame precision issue

+++ maro/backends/raw/common.h
@@ -29,7 +29,7 @@ namespace maro
 
       using NODE_INDEX = uint32_t;
       using SLOT_INDEX = uint32_t;
-      using QUERY_FLOAT = float;
+      using QUERY_FLOAT = double;
 
       using ATTR_CHAR = char;
       using ATTR_UCHAR = unsigned char;
